updating the aleatory uncertainty
The uncertaintyPropagation needs more work and the attention of Leslie. the outcome of the sampling methods if all input are distributions is K-S bounds, how to we wrap them into an uncertain number? I also need to create an uncertainty_propagation function which will account for epistemic if all input is intervals to aleatory to mixed...